[1.19.2] Revision of ITooltipProvider system.

+++ src/main/java/org/zeith/hammerlib/tiles/tooltip/own/IRenderableInfo.java
@@ -6,10 +6,10 @@ import net.minecraftforge.api.distmarker.OnlyIn;
 
 public interface IRenderableInfo
 {
-	int getWidth();
-
-	int getHeight();
-
+	float getWidth();
+	
+	float getHeight();
+	
 	@OnlyIn(Dist.CLIENT)
 	void render(PoseStack matrix, float x, float y, float partialTime);
 }
